# libs/bits.BitArray: memoize and reuse trueIndices slice instead of discarding always