Noticed looking at 10s of Tharsis CPU and RAM profiles that
libs/bits.BitArray.getTrueIndices consumed a considerable amount of RAM,
on reading the code, I noticed that getTrueIndices is ONLY invoked
by PickRandom, thus is was wastefully allocating and then discarding
the slice being used, derived from .Bits which NEVER changes unless
.FromProto is invoked.

This change memoizes the slice for reuse and if .Bits changes, thus
the memoized slice's capacity changes, we'll refresh the slice.

The results from this change show a reduction in ~10% of overall RAM for
the program, and then some also in CPU time:
```shell
$ benchstat before.txt after.txt
name                  old time/op    new time/op    delta
BitArrayPickRandom-8    5.15µs ± 5%    4.62µs ± 3%   -10.15%  (p=0.000 n=20+19)

name                  old alloc/op   new alloc/op   delta
BitArrayPickRandom-8    2.16kB ± 0%    0.00kB       -100.00%  (p=0.000 n=20+20)

name                  old allocs/op  new allocs/op  delta
BitArrayPickRandom-8      6.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
```

Fixes #7419

Please add a description of the changes that this PR introduces and the files that
are the most critical to review.

If this PR fixes an open Issue, please include "Closes #XXX" (where "XXX" is the Issue number) 
so that GitHub will automatically close the Issue when this PR is merged.